+++ src/main/java/com/irurueta/numerical/fitting/LevenbergMarquardtMultiVariateFitter.java
@@ -750,10 +750,10 @@ public class LevenbergMarquardtMultiVariateFitter extends MultiVariateFitter {
                 }
 
                 // add to mse
-                mse += dy * dy / (double) Math.abs(degreesOfFreedom);
+                mse += dy * dy / Math.abs(degreesOfFreedom);
 
                 // and find chi square
-                chisq += dy * dy * sig2i / (double) degreesOfFreedom;
+                chisq += dy * dy * sig2i / degreesOfFreedom;
             }
         }
